Complete tenant comparison implementation with system test
Implements metadata comparison between tenants with proper DI registration:
- Add MetadataComparator to compare tenant metadata and identify differences
- Add MetadataLoader to load tenant metadata from MongoDB with optional comparison areas
- Update TenantComparisonService to use loader and comparator for metadata comparison
- Fix DI registration in AddTenantComparison to include MetadataLoader and MetadataComparator
- Update ServiceCollectionFixture to register tenant comparison services
- Add comprehensive TenantComparisonServiceTests with two tenant comparison test
- Fix transaction initialization in MetadataLoader (add StartTransaction call)
- Change TenantMetadata.CkModelCount from int to long for consistency

The test creates two tenants, compares them using ITenantComparisonService from DI,
and verifies the comparison report is generated correctly with metadata differences.

+++ src/Runtime.Engine.MongoDb/Comparison/Models/TenantMetadata.cs
@@ -33,5 +33,5 @@ public class TenantMetadata
     /// <summary>
     ///     Number of imported CkModels
     /// </summary>
-    public int CkModelCount { get; set; }
+    public long CkModelCount { get; set; }
 }
